% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS.R
\name{addNoAnswer}
\alias{addNoAnswer}
\title{add levels to recommendation factor}
\usage{
addNoAnswer(x)
}
\arguments{
\item{x}{vector numeric}
}
\value{
factor
}
\description{
add levels to recommendation factor
}
\examples{
addNoAnswer(c(1,2))
}
