% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_my_dataset.R
\docType{data}
\name{my_dataset}
\alias{my_dataset}
\title{Recommendation data}
\format{
data.frame
}
\source{
factice
}
\usage{
my_dataset
}
\description{
Un jeu de données contenant des données factices de satisfaction
}
\author{
Colin \email{nmanoharan@me.com}
}
\keyword{data}
