% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS_decompose.R
\name{NPS_decompose}
\alias{NPS_decompose}
\title{NPS_decompose}
\usage{
NPS_decompose(reco)
}
\arguments{
\item{reco}{vector}
}
\value{
data.frame
}
\description{
NPS_decompose
}
\examples{
NPS_decompose(my_dataset$reco)
}
