% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS_evolution.R
\name{NPS_evolution_plot}
\alias{NPS_evolution_plot}
\title{Plot evolution of NPS}
\usage{
NPS_evolution_plot(reco, date, unite)
}
\arguments{
\item{reco}{vector numeric}

\item{date}{vector date (dmy)}

\item{unite}{character (mois, jours, annee)}
}
\value{
plotly
}
\description{
Plot evolution of NPS
}
\examples{
NPS_evolution_plot(c(1,10),c("01/01/2020","01/02/2020"),"mois")
}
