% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS_evolution_repartition.R
\name{NPS_evolution_repartition}
\alias{NPS_evolution_repartition}
\title{Evolution de la repartition du NPS}
\usage{
NPS_evolution_repartition(reco, date, unite)
}
\arguments{
\item{reco}{vector}

\item{date}{vector (dmy)}

\item{unite}{(mois,annee,jours)}
}
\value{
plotly object
}
\description{
Evolution de la repartition du NPS
}
\examples{
NPS_evolution_repartition(my_dataset$reco,my_dataset$date,"mois")
}
