% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPS_hbarplot.R
\name{NPS_hbarplot}
\alias{NPS_hbarplot}
\title{NPS_hbarplot}
\usage{
NPS_hbarplot(dataset, dr, reco)
}
\arguments{
\item{dataset}{data.frame}

\item{dr}{character}

\item{reco}{numeric}
}
\value{
plotlyobject
}
\description{
NPS_hbarplot
}
\examples{
NPS_hbarplot(my_dataset,dr, reco)
}
