% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sat_dognuts.R
\name{Sat_Dognuts}
\alias{Sat_Dognuts}
\title{Generate Plotly dognuts about satisfaction question}
\usage{
Sat_Dognuts(attribut, nb_level, titre)
}
\arguments{
\item{attribut}{vector}

\item{nb_level}{numeric}

\item{titre}{character}
}
\value{
plotly
}
\description{
Generate Plotly dognuts about satisfaction question
}
\examples{
Sat_Dognuts(my_dataset$reco,10,titre = "Propreté de l'accueil")
}
